+++ react.client/CHANGELOG.md
@@ -4,11 +4,11 @@ The following tools were used to generate this project:
 - create-vite
 
 The following steps were used to generate this project:
-- Create react project with create-vite: `npm init --yes vite@latest reactfullstackts.client -- --template=react-ts`.
+- Create react project with create-vite: `npm init --yes vite@latest react.client -- --template=react-ts`.
 - Update `vite.config.ts` to set up proxying and certs.
 - Add `@type/node` for `vite.config.js` typing.
 - Update `App` component to fetch and display weather information.
-- Create project file (`reactfullstackts.client.esproj`).
+- Create project file (`react.client.esproj`).
 - Create `launch.json` to enable debugging.
 - Create `tasks.json` to enable debugging.
 - Add project to solution.